DotFileFormat: Abstract out report format
Reimplement nested report format as DotFileFormatNested.
