+++ src/main/java/net/kemitix/dependency/digraph/maven/plugin/ReportGenerator.java
@@ -1,7 +1,5 @@
 package net.kemitix.dependency.digraph.maven.plugin;
 
-import net.kemitix.node.Node;
-
 /**
  * Interface for creating dependency data report.
  *
@@ -12,10 +10,10 @@ interface ReportGenerator extends MojoService {
     /**
      * Generates the dependency report.
      *
-     * @param baseNode the node to report on
+     * @param dotFileFormat the format specific report generator
      *
      * @return the report
      */
-    String generate(Node<PackageData> baseNode);
+    String generate(DotFileFormat dotFileFormat);
 
 }
